requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
razorpay
python-dotenv